DAY06PythonII-1-almondsm/src/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pyTelegramBotAPI==4.14.0
python-dotenv==0.21.1
requests==2.31.0
telebot==0.0.5
urllib3==2.0.7
